% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TNC.R
\docType{package}
\name{TNC}
\alias{TNC}
\alias{TNC-package}
\title{TNC: A package for computing temporal network centrality values for nodes of
a dynamic network.}
\description{
The TNC package provides three functions: tbc, tcc and tdc for calculating
temporal betweenness, temporal closeness and temporal degree centrality.
}
